Refactored interfaces
