+++ src/@interfaces/IColumn.ts
@@ -1,0 +1,5 @@
+import { BoxProps } from "@chakra-ui/react"
+
+export interface IColumn extends BoxProps {
+  col?: string[]
+}
