+++ src/@interfaces/IGrid.ts
@@ -1,0 +1,11 @@
+export interface IGrid {
+  sm: {
+    [key: string]: string
+  }
+  md: {
+    [key: string]: string
+  }
+  lg: {
+    [key: string]: string
+  }
+}
