+++ src/@interfaces/IHamburgerButton.ts
@@ -1,0 +1,8 @@
+import { BoxProps } from "@chakra-ui/react"
+
+export interface IHamburgerButton extends BoxProps {
+  data: {
+    ariaLabel: string
+    menuLabel: string
+  }
+}
