+++ src/@interfaces/ILink.ts
@@ -1,0 +1,7 @@
+import { LinkProps } from "@chakra-ui/react"
+
+export interface ILink extends LinkProps {
+  asButton?: boolean
+  sizeButton?: string
+  href: string
+}
