+++ src/@interfaces/IText.ts
@@ -1,0 +1,17 @@
+import { TextProps } from "@chakra-ui/react"
+import { HTMLMotionProps } from "framer-motion"
+
+type FramerMotionProps =
+  | HTMLMotionProps<"h1">
+  | HTMLMotionProps<"h2">
+  | HTMLMotionProps<"h3">
+  | HTMLMotionProps<"p">
+type ChakraTextProps = TextProps
+
+type CustomTextProps = {
+  type: string
+}
+
+export type ChakraTextWithMotionType = CustomTextProps &
+  ChakraTextProps &
+  FramerMotionProps
