+++ src/@interfaces/index.ts
@@ -1,0 +1,5 @@
+export * from "./IHamburgerButton"
+export * from "./IGrid"
+export * from "./IColumn"
+export * from "./IText"
+export * from "./ILink"
