+++ src/components/layout/Row.tsx
@@ -1,11 +1,7 @@
 import React from "react"
 import { Flex, FlexOptions } from "@chakra-ui/react"
 
-interface IRow extends FlexOptions {
-  children?: React.ReactNode
-}
-
-export const Row: React.FC<IRow> = ({ children, ...props }) => {
+export const Row: React.FC<FlexOptions> = ({ children, ...props }) => {
   return (
     <Flex className="row" {...props}>
       {children}
